+++ src/com/cjimenezro/Electricidad_Chispas/data/CustomerDataStore.java
@@ -1,6 +1,8 @@
 package com.cjimenezro.Electricidad_Chispas.data;
 
-import com.cjimenezro.Electricidad_Chispas.Clientes;
+import com.cjimenezro.Electricidad_Chispas.domain.models.Clientes;
+
+import java.util.List;
 
 public interface CustomerDataStore {
 
@@ -9,6 +11,9 @@ public interface CustomerDataStore {
     void deleteCustomer(Clientes cliente);
 
     void updateCustomer(Clientes cliente);
+
+    List <Clientes> getAllCustomers();
+    Clientes findById (Integer codCliente);
