+++ src/com/cjimenezro/Electricidad_Chispas/data/MemCustomerDataStore.java
@@ -1,10 +1,14 @@
 package com.cjimenezro.Electricidad_Chispas.data;
 
-import com.cjimenezro.Electricidad_Chispas.Clientes;
+import com.cjimenezro.Electricidad_Chispas.domain.models.Clientes;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.TreeMap;
 
 public class MemCustomerDataStore implements CustomerDataStore{
 
-    private
+    private TreeMap<Integer, Clientes> dataStore = new TreeMap<>();
 
     @Override
     public void saveCustomer(Clientes cliente) {
@@ -19,5 +23,15 @@ public class MemCustomerDataStore implements CustomerDataStore{
     @Override
     public void updateCustomer(Clientes cliente) {
         dataStore.replace(cliente.getCodigoCliente(),cliente);
+    }
+
+    @Override
+    public List<Clientes> getAllCustomers() {
+        return new ArrayList<>(dataStore.values());
+    }
+
+    @Override
+    public Clientes findById(Integer codCliente) {
+        return null;
     }
 }
