+++ src/com/cjimenezro/Electricidad_Chispas/domain/models/Clientes.java
@@ -1,4 +1,4 @@
-package com.cjimenezro.Electricidad_Chispas;
+package com.cjimenezro.Electricidad_Chispas.domain.models;
 
 public interface Clientes {
