+++ src/com/cjimenezro/Electricidad_Chispas/domain/models/Facturas.java
@@ -1,4 +1,4 @@
-package com.cjimenezro.Electricidad_Chispas;
+package com.cjimenezro.Electricidad_Chispas.domain.models;
 
 public class Facturas {
     private int codigoFactura,baseImponible,total;
