+++ src/com/cjimenezro/Electricidad_Chispas/domain/models/ProductosServicios.java
@@ -1,4 +1,4 @@
-package com.cjimenezro.Electricidad_Chispas;
+package com.cjimenezro.Electricidad_Chispas.domain.models;
 
 public interface ProductosServicios {
